requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
requests==2.32.3
python-dotenv==1.0.1
groq==0.9.0
motor==3.4.0
httpx==0.27.0
pymongo[srv]==4.7.2
python-multipart==0.0.9
twilio==9.0.4
